UPDATE: all optput mqtt classes to new format
